feat: Enhance Book and Order models for mass assignment and add new fields
feat: Add status field to Order model and implement scopes for order status

fix: Update My Books migration to prevent duplicate purchases by the same user

fix: Change total_price field type in Orders migration to decimal for accuracy

fix: Update price field type in Order Items migration to decimal for accuracy

feat: Seed database with sample data including users, books, orders, and ratings

feat: Implement new API routes for book search, best sellers, and newest books

feat: Add user profile management routes and functionality

feat: Create FavoriteController for managing user favorites

feat: Implement RatingController for book ratings and reviews

feat: Add StatsController for admin statistics and analytics

feat: Create UserController for admin user management

feat: Add request validation for favorites and ratings

+++ database/migrations/2026_05_15_162635_create_order_items_table.php
@@ -14,7 +14,8 @@ return new class extends Migration {
             $table->id();
             $table->foreignId('order_id')->constrained()->onDelete('cascade');
             $table->foreignId('book_id')->constrained()->onDelete('cascade');
-            $table->integer('price');
+            // ✅ إصلاح: integer → decimal(10,2)
+            $table->decimal('price', 10, 2);
             $table->timestamps();
         });
     }
